필터 쿼리스트링 통신 완료
