+++ src/Components/Nav/Nav.js
@@ -1,12 +1,12 @@
 import React, { useState, useEffect } from 'react';
 import styled, { css } from 'styled-components';
 import { useSelector } from 'react-redux';
+import { useHistory } from 'react-router-dom';
 import NavList from './NavList';
 import { KeyboardArrowLeft } from '@styled-icons/material';
-// import SELLER_NAV from './DataSellerNav';
-// import MASTER_NAV from './DataMasterNav';
 
 export default function Nav() {
+  const history = useHistory();
   const [active, setActive] = useState([0, 0]);
   const [navData, setNavData] = useState([]);
   const [subActive, setSubActive] = useState(0);
@@ -15,22 +15,22 @@ export default function Nav() {
   // store에 있는 nav 정보를 가져온다.
   const { nav_list, filter_list } = useSelector(({ userInfo }) => ({
     nav_list: userInfo.nav_list,
-    filter_list: userInfo.filter_list,
   }));
 
   //nav가 변경되면 nav 데이터를 최신화
   useEffect(() => {
+    if (!nav_list[0]) {
+      alert('다시 로그인 해주세요.');
+      history.push('/');
+    }
+
     setNavData(nav_list);
-    console.log(nav_list);
-    console.log('123123123123123', filter_list);
   }, [nav_list]);
 
   // 네브가 눌렸을때, 페이지 이동 함수
   const handlePage = (menuTitle, subTitle) => {
     setActive([menuTitle, subTitle]);
   };
-
-  console.log(nav_list);
 
   return (
     <NavContainer sidebarSmall={sidebarSmall}>
@@ -62,6 +62,9 @@ export default function Nav() {
 }
 
 const NavContainer = styled.div`
+  @media only screen and (max-width: 934px) {
+    display: none;
+  }
   width: 215px;
   height: 100vh;
   padding-top: 45px;
@@ -72,9 +75,6 @@ const NavContainer = styled.div`
     css`
       width: 40px;
     `}
-  @media only screen and (max-width: 934px) {
-    display: none;
-  }
 `;
 
 const NavBox = styled.ul`
