+++ src/Routes.js
@@ -7,7 +7,6 @@ import ProductManagement from './Pages/Product/ProductManagement';
 import Home from './Pages/Home/Home';
 import Order from './Pages/Order/Order';
 import User from './Pages/User/User';
-import UserDetail from './Pages/User/UserDetail';
 
 function Routes() {
   return (
@@ -18,10 +17,10 @@ function Routes() {
         <Route exact path='/main' component={Main} />
         <Route exact path='/home' component={Home} />
         <Route exact path='/order/:id' component={Order} />
-        {/* <Route exact path='/account/seller' component={User} /> */}
-        <Route exact path='/account/sellerList' component={User} />
+        <Route exact path='/account/seller' component={User} />
+        {/* <Route exact path='/account/sellerlist' component={User} /> */}
+        {/* <Route exact path='/user' component={User} /> */}
         <Route exact path='/product' component={ProductManagement} />
-        <Route exact path='/userdetail' componenet={UserDetail} />
       </Switch>
     </Router>
   );
