+++ src/config.js
@@ -1,1 +1,1 @@
-export const API = 'http://10.58.7.141:';
+export const API = 'http://192.168.7.14:';
